pyproject.toml
[project]
name = "alfa"
version = "0.1.0"
description = "Math Problem Explanation App with Gradio and LLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=4.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]
